requirements.txt
asgiref==3.8.1
bcrypt==4.2.0
bidict==0.23.1
blinker==1.8.2
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.7
colorama==0.4.6
contourpy==1.3.1
crispy-bootstrap5==2024.10
cryptography==44.0.2
cycler==0.12.1
defusedxml==0.7.1
distlib==0.3.9
dj-database-url==3.0.0
Django==5.1.2
django-allauth==65.6.0
django-crispy-forms==2.3
django-filter==25.1
djangorestframework==3.15.2
dnspython==1.16.0
docutils==0.21.2
eventlet==0.31.0
filelock==3.16.1
filetype==1.2.0
Flask==3.0.3
Flask-Bcrypt==1.0.1
Flask-Login==0.6.3
Flask-SocketIO==5.1.1
Flask-SQLAlchemy==3.1.1
fonttools==4.56.0
greenlet==3.1.1
gunicorn==23.0.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
Kivy==2.3.1
kivy-deps.angle==0.4.0
kivy-deps.glew==0.3.1
Kivy-Garden==0.1.5
kivy_deps.sdl2==0.8.0
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.1
mlxtend==0.23.4
numpy==2.2.4
oauthlib==3.2.2
packaging==24.2
pandas==2.2.3
pillow==11.0.0
platformdirs==4.3.6
psycopg2-binary==2.9.10
pycparser==2.22
pygame==2.6.1
Pygments==2.19.1
PyJWT==2.10.1
pyparsing==3.2.1
pypiwin32==223
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-engineio==4.2.0
python-socketio==5.3.0
python3-openid==3.2.0
pytz==2025.1
pywin32==306
requests==2.32.3
requests-oauthlib==2.0.0
scikit-learn==1.6.1
scipy==1.15.2
seaborn==0.13.2
six==1.16.0
social-auth-app-django==5.4.3
social-auth-core==4.5.6
SQLAlchemy==2.0.36
sqlparse==0.5.1
threadpoolctl==3.6.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
virtualenv==20.27.1
Werkzeug==3.0.4
whitenoise==6.9.0
